feat(owner-tool): add an option to export OVs from the Manufacturer database
This adds the `export_manufacturer_vouchers` option to the Owner CLI, which
allows to export OVs from the Manufacturer Server's DB to a directory. The
options are to export a single OV by GUID or all of them.

Updated the DB trait to add the `select *` query that yields all the OVs in the
Manufacturer.

Signed-off-by: Irene Diez <idiez@redhat.com>